# openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /kratosblog/{name}:
        get:
            tags:
                - Greeter
            description: Sends a greeting
            operationId: Greeter_SayHello
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HelloReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/articles/category:
        get:
            tags:
                - Article
            operationId: Article_CategoryList
            parameters:
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CategoryListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/articles/detail:
        get:
            tags:
                - Article
            operationId: Article_Detail
            parameters:
                - name: id
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DetailReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/articles/list:
        get:
            tags:
                - Article
            operationId: Article_List
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/articles/tag:
        get:
            tags:
                - Article
            operationId: Article_TagList
            parameters:
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TagListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/category/list:
        get:
            tags:
                - Category
            operationId: Category_ListCategory
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListCategoryReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        Article:
            type: object
            properties:
                id:
                    type: string
                title:
                    type: string
                cover:
                    $ref: '#/components/schemas/Image'
                content:
                    type: string
                tags:
                    type: array
                    items:
                        $ref: '#/components/schemas/Tag'
                category:
                    $ref: '#/components/schemas/Category'
                updatedAt:
                    type: string
                    format: date-time
        Category:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                cover:
                    $ref: '#/components/schemas/Image'
                description:
                    type: string
                color:
                    type: string
                articleCount:
                    type: string
        CategoryListReply:
            type: object
            properties:
                total:
                    type: integer
                    format: int64
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/Article'
        DetailReply:
            type: object
            properties:
                detail:
                    $ref: '#/components/schemas/Article'
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        HelloReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the greetings
        Image:
            type: object
            properties:
                url:
                    type: string
                color:
                    type: string
        ListCategoryReply:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/Category'
        ListReply:
            type: object
            properties:
                total:
                    type: integer
                    format: int64
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/Article'
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        Tag:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                color:
                    type: string
                articleCount:
                    type: string
        TagListReply:
            type: object
            properties:
                total:
                    type: integer
                    format: int64
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/Article'
tags:
    - name: Article
    - name: Category
    - name: Greeter
      description: The greeting service definition.